Fix error in vpp_intf_convert_one_if()
It was referencing a vector element by using an
interface sw_if_index as an offset. This was from the
old libvppmgmt API call to retrieve interface data. The
newer call just passes back the information for the one
interface and the if index doesn't need to be used to
find the right value.

It's by pure luck that this worked at all for any
interfaces.